fix: remove cross-branch contamination and fix http-cache bugs
- Remove files accidentally committed from PRMonitor decomposition branch
- Fix double-wrapped Promise in cachedRequest dedup path (await existing)
- Re-read cache in 304 handler to avoid stale closure snapshot
- Set explicit 0o600 mode on cache file writes
- Fix prettier formatting in index.ts
