+++ src/core/pr-monitor.test.ts
@@ -29,12 +29,6 @@ const {
 const { isAcknowledgmentComment } = await import('./comment-utils.js');
 const { getStateManager } = await import('./state.js');
 
-// Import extracted module functions (pure logic, no mocks needed)
-const { analyzeChecklist } = await import('./checklist-analysis.js');
-const { extractMaintainerActionHints } = await import('./maintainer-analysis.js');
-const { determineReviewDecision, getLatestChangesRequestedDate, checkUnrespondedComments, isAllSelfReplies } =
-  await import('./review-analysis.js');
-
 describe('PRMonitor CI status deduplication', () => {
   const emptyCombinedStatus = {
     data: {
@@ -472,84 +466,94 @@ describe('PRMonitor analyzeChecklist', () => {
   });
 
   it('should return false for empty body', () => {
-    const result = analyzeChecklist('');
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).analyzeChecklist('');
     expect(result.hasIncompleteChecklist).toBe(false);
     expect(result.checklistStats).toBeUndefined();
   });
 
   it('should return false for body with no checkboxes', () => {
-    const result = analyzeChecklist('This PR adds a new feature.');
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).analyzeChecklist('This PR adds a new feature.');
     expect(result.hasIncompleteChecklist).toBe(false);
   });
 
   it('should return false when all checkboxes are checked', () => {
+    const monitor = new PRMonitor('fake-token');
     const body = '- [x] Tests added\n- [x] Docs updated\n- [x] Linted';
-    const result = analyzeChecklist(body);
+    const result = (monitor as any).analyzeChecklist(body);
     expect(result.hasIncompleteChecklist).toBe(false);
     expect(result.checklistStats).toEqual({ checked: 3, total: 3 });
   });
 
   it('should return true when some checkboxes are unchecked', () => {
+    const monitor = new PRMonitor('fake-token');
     const body = '- [x] Tests added\n- [ ] Docs updated\n- [x] Linted';
-    const result = analyzeChecklist(body);
+    const result = (monitor as any).analyzeChecklist(body);
     expect(result.hasIncompleteChecklist).toBe(true);
     expect(result.checklistStats).toEqual({ checked: 2, total: 3 });
   });
 
   it('should return true when all checkboxes are unchecked', () => {
+    const monitor = new PRMonitor('fake-token');
     const body = '- [ ] Tests added\n- [ ] Docs updated';
-    const result = analyzeChecklist(body);
+    const result = (monitor as any).analyzeChecklist(body);
     expect(result.hasIncompleteChecklist).toBe(true);
     expect(result.checklistStats).toEqual({ checked: 0, total: 2 });
   });
 
   it('should handle case-insensitive checked marks (- [X])', () => {
+    const monitor = new PRMonitor('fake-token');
     const body = '- [X] Done\n- [ ] Not done';
-    const result = analyzeChecklist(body);
+    const result = (monitor as any).analyzeChecklist(body);
     expect(result.hasIncompleteChecklist).toBe(true);
     expect(result.checklistStats).toEqual({ checked: 1, total: 2 });
   });
 
   it('should not flag conditional checklist items as incomplete (#152)', () => {
+    const monitor = new PRMonitor('fake-token');
     const body = [
       '- [x] PR title and summary are descriptive',
       '- [x] Docs updated or follow-up ticket created',
       '- [x] Tests included',
       '- [ ] PR Labeled with `release/backport` (if the PR is an urgent fix that needs to be backported)',
     ].join('\n');
-    const result = analyzeChecklist(body);
+    const result = (monitor as any).analyzeChecklist(body);
     expect(result.hasIncompleteChecklist).toBe(false);
     expect(result.checklistStats).toEqual({ checked: 3, total: 4 });
   });
 
   it('should still flag non-conditional unchecked items alongside conditional ones (#152)', () => {
+    const monitor = new PRMonitor('fake-token');
     const body = ['- [x] Tests included', '- [ ] Documentation updated', '- [ ] Changelog entry (if applicable)'].join(
       '\n',
     );
-    const result = analyzeChecklist(body);
+    const result = (monitor as any).analyzeChecklist(body);
     expect(result.hasIncompleteChecklist).toBe(true);
     expect(result.checklistStats).toEqual({ checked: 1, total: 3 });
   });
 
   it('should handle multiple conditional patterns (#152)', () => {
+    const monitor = new PRMonitor('fake-token');
     const body = [
       '- [x] Tests pass',
       '- [ ] Update CHANGELOG (if applicable)',
       '- [ ] Add migration (only if schema changes)',
       '- [ ] Backport label (if needed)',
     ].join('\n');
-    const result = analyzeChecklist(body);
+    const result = (monitor as any).analyzeChecklist(body);
     expect(result.hasIncompleteChecklist).toBe(false);
     expect(result.checklistStats).toEqual({ checked: 1, total: 4 });
   });
 
   it('should handle "N/A" and "optional" patterns (#152)', () => {
+    const monitor = new PRMonitor('fake-token');
     const body = [
       '- [x] Code review',
       '- [ ] Screenshots (N/A for backend changes)',
       '- [ ] Optional: update README',
     ].join('\n');
-    const result = analyzeChecklist(body);
+    const result = (monitor as any).analyzeChecklist(body);
     expect(result.hasIncompleteChecklist).toBe(false);
     expect(result.checklistStats).toEqual({ checked: 1, total: 3 });
   });
@@ -561,27 +565,38 @@ describe('PRMonitor extractMaintainerActionHints', () => {
   });
 
   it('should return changes_requested when reviewDecision is changes_requested', () => {
-    const result = extractMaintainerActionHints(undefined, 'changes_requested');
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).extractMaintainerActionHints(undefined, 'changes_requested');
     expect(result).toEqual(['changes_requested']);
   });
 
   it('should return empty array when no comment and no changes_requested', () => {
-    const result = extractMaintainerActionHints(undefined, 'approved');
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).extractMaintainerActionHints(undefined, 'approved');
     expect(result).toEqual([]);
   });
 
   it('should detect demo_requested keywords', () => {
-    const result = extractMaintainerActionHints('Can you show a screenshot of the before/after?', 'review_required');
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).extractMaintainerActionHints(
+      'Can you show a screenshot of the before/after?',
+      'review_required',
+    );
     expect(result).toContain('demo_requested');
   });
 
   it('should detect tests_requested keywords', () => {
-    const result = extractMaintainerActionHints('Please add test coverage for this feature', 'review_required');
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).extractMaintainerActionHints(
+      'Please add test coverage for this feature',
+      'review_required',
+    );
     expect(result).toContain('tests_requested');
   });
 
   it('should detect docs_requested keywords', () => {
-    const result = extractMaintainerActionHints(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).extractMaintainerActionHints(
       'Can you update the documentation for this API change?',
       'review_required',
     );
@@ -589,12 +604,17 @@ describe('PRMonitor extractMaintainerActionHints', () => {
   });
 
   it('should detect rebase_requested keywords', () => {
-    const result = extractMaintainerActionHints('This branch is behind main, could you rebase?', 'review_required');
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).extractMaintainerActionHints(
+      'This branch is behind main, could you rebase?',
+      'review_required',
+    );
     expect(result).toContain('rebase_requested');
   });
 
   it('should detect multiple hints in one comment', () => {
-    const result = extractMaintainerActionHints(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).extractMaintainerActionHints(
       'Please add test coverage and a screenshot of the changes',
       'changes_requested',
     );
@@ -610,22 +630,28 @@ describe('PRMonitor determineReviewDecision', () => {
   });
 
   it('should return review_required when no reviews exist', () => {
-    const result = determineReviewDecision([]);
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).determineReviewDecision([]);
     expect(result).toBe('review_required');
   });
 
   it('should return approved when latest review is APPROVED', () => {
-    const result = determineReviewDecision([{ state: 'APPROVED', user: { login: 'reviewer1' } }]);
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).determineReviewDecision([{ state: 'APPROVED', user: { login: 'reviewer1' } }]);
     expect(result).toBe('approved');
   });
 
   it('should return changes_requested when latest review is CHANGES_REQUESTED', () => {
-    const result = determineReviewDecision([{ state: 'CHANGES_REQUESTED', user: { login: 'reviewer1' } }]);
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).determineReviewDecision([
+      { state: 'CHANGES_REQUESTED', user: { login: 'reviewer1' } },
+    ]);
     expect(result).toBe('changes_requested');
   });
 
   it('should use latest review per user', () => {
-    const result = determineReviewDecision([
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).determineReviewDecision([
       { state: 'CHANGES_REQUESTED', user: { login: 'reviewer1' } },
       { state: 'APPROVED', user: { login: 'reviewer1' } }, // Same user, later approval
     ]);
@@ -633,7 +659,8 @@ describe('PRMonitor determineReviewDecision', () => {
   });
 
   it('should prioritize changes_requested over approved from different users', () => {
-    const result = determineReviewDecision([
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).determineReviewDecision([
       { state: 'APPROVED', user: { login: 'reviewer1' } },
       { state: 'CHANGES_REQUESTED', user: { login: 'reviewer2' } },
     ]);
@@ -641,7 +668,8 @@ describe('PRMonitor determineReviewDecision', () => {
   });
 
   it('should return review_required for COMMENTED-only reviews', () => {
-    const result = determineReviewDecision([{ state: 'COMMENTED', user: { login: 'reviewer1' } }]);
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).determineReviewDecision([{ state: 'COMMENTED', user: { login: 'reviewer1' } }]);
     expect(result).toBe('review_required');
   });
 });
@@ -652,24 +680,30 @@ describe('PRMonitor getLatestChangesRequestedDate', () => {
   });
 
   it('should return undefined when no reviews', () => {
-    const result = getLatestChangesRequestedDate([]);
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).getLatestChangesRequestedDate([]);
     expect(result).toBeUndefined();
   });
 
   it('should return undefined when no CHANGES_REQUESTED reviews', () => {
-    const result = getLatestChangesRequestedDate([{ state: 'APPROVED', submitted_at: '2026-02-07T10:00:00Z' }]);
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).getLatestChangesRequestedDate([
+      { state: 'APPROVED', submitted_at: '2026-02-07T10:00:00Z' },
+    ]);
     expect(result).toBeUndefined();
   });
 
   it('should return the date of the single CHANGES_REQUESTED review', () => {
-    const result = getLatestChangesRequestedDate([
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).getLatestChangesRequestedDate([
       { state: 'CHANGES_REQUESTED', submitted_at: '2026-02-07T10:00:00Z' },
     ]);
     expect(result).toBe('2026-02-07T10:00:00Z');
   });
 
   it('should return the latest date when multiple CHANGES_REQUESTED reviews', () => {
-    const result = getLatestChangesRequestedDate([
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).getLatestChangesRequestedDate([
       { state: 'CHANGES_REQUESTED', submitted_at: '2026-02-05T10:00:00Z' },
       { state: 'APPROVED', submitted_at: '2026-02-06T10:00:00Z' },
       { state: 'CHANGES_REQUESTED', submitted_at: '2026-02-08T10:00:00Z' },
@@ -710,13 +744,15 @@ describe('PRMonitor checkUnrespondedComments', () => {
   });
 
   it('should return false when no comments or reviews', () => {
-    const result = checkUnrespondedComments([], [], [], 'testuser');
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments([], [], [], 'testuser');
     expect(result.hasUnrespondedComment).toBe(false);
     expect(result.lastMaintainerComment).toBeUndefined();
   });
 
   it('should return false when only user comments exist', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [{ user: { login: 'testuser' }, body: 'My comment', created_at: '2026-02-07T10:00:00Z' }],
       [],
       [],
@@ -726,7 +762,8 @@ describe('PRMonitor checkUnrespondedComments', () => {
   });
 
   it('should return true when maintainer commented after user', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [
         { user: { login: 'testuser' }, body: 'My PR', created_at: '2026-02-07T10:00:00Z' },
         { user: { login: 'maintainer' }, body: 'Please fix X', created_at: '2026-02-07T12:00:00Z' },
@@ -740,7 +777,8 @@ describe('PRMonitor checkUnrespondedComments', () => {
   });
 
   it('should return false when user replied after maintainer', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [
         { user: { login: 'maintainer' }, body: 'Please fix X', created_at: '2026-02-07T10:00:00Z' },
         { user: { login: 'testuser' }, body: 'Fixed', created_at: '2026-02-07T12:00:00Z' },
@@ -753,7 +791,8 @@ describe('PRMonitor checkUnrespondedComments', () => {
   });
 
   it('should skip bot comments with [bot] suffix', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [
         { user: { login: 'testuser' }, body: 'My PR', created_at: '2026-02-07T10:00:00Z' },
         { user: { login: 'dependabot[bot]' }, body: 'Dependency update', created_at: '2026-02-07T12:00:00Z' },
@@ -766,9 +805,10 @@ describe('PRMonitor checkUnrespondedComments', () => {
   });
 
   it('should skip known bot accounts without [bot] suffix', () => {
+    const monitor = new PRMonitor('fake-token');
     const knownBots = ['CLAassistant', 'codecov-commenter', 'changeset-bot', 'netlify', 'sonarcloud'];
     for (const bot of knownBots) {
-      const result = checkUnrespondedComments(
+      const result = (monitor as any).checkUnrespondedComments(
         [
           { user: { login: 'testuser' }, body: 'My PR', created_at: '2026-02-07T10:00:00Z' },
           { user: { login: bot }, body: 'Automated check passed', created_at: '2026-02-07T12:00:00Z' },
@@ -782,7 +822,8 @@ describe('PRMonitor checkUnrespondedComments', () => {
   });
 
   it('should still flag non-bot maintainer comments when bot comments also exist', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [
         { user: { login: 'testuser' }, body: 'My PR', created_at: '2026-02-07T10:00:00Z' },
         { user: { login: 'CLAassistant' }, body: 'CLA signed', created_at: '2026-02-07T11:00:00Z' },
@@ -797,7 +838,8 @@ describe('PRMonitor checkUnrespondedComments', () => {
   });
 
   it('should skip comments from deleted accounts (null user)', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [
         { user: { login: 'testuser' }, body: 'My PR', created_at: '2026-02-07T10:00:00Z' },
         { user: null, body: 'Ghost comment', created_at: '2026-02-07T12:00:00Z' },
@@ -810,7 +852,8 @@ describe('PRMonitor checkUnrespondedComments', () => {
   });
 
   it('should include review comments with body text in timeline', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [],
       [{ user: { login: 'maintainer' }, body: 'Needs changes here', submitted_at: '2026-02-07T12:00:00Z' }],
       [],
@@ -821,7 +864,8 @@ describe('PRMonitor checkUnrespondedComments', () => {
   });
 
   it('should skip reviews with empty body', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [],
       [{ user: { login: 'maintainer' }, body: '', submitted_at: '2026-02-07T12:00:00Z' }],
       [],
@@ -831,8 +875,9 @@ describe('PRMonitor checkUnrespondedComments', () => {
   });
 
   it('should truncate long comment bodies to 200 chars', () => {
+    const monitor = new PRMonitor('fake-token');
     const longBody = 'x'.repeat(300);
-    const result = checkUnrespondedComments(
+    const result = (monitor as any).checkUnrespondedComments(
       [{ user: { login: 'maintainer' }, body: longBody, created_at: '2026-02-07T12:00:00Z' }],
       [],
       [],
@@ -842,7 +887,8 @@ describe('PRMonitor checkUnrespondedComments', () => {
   });
 
   it('should be case-insensitive for username matching', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [
         { user: { login: 'TestUser' }, body: 'My PR', created_at: '2026-02-07T10:00:00Z' },
         { user: { login: 'maintainer' }, body: 'Fix this', created_at: '2026-02-07T12:00:00Z' },
@@ -1480,7 +1526,8 @@ describe('isAcknowledgmentComment (Issue #69)', () => {
   });
 
   it('should not trigger hasUnrespondedComment for acknowledgment comments', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [
         { user: { login: 'testuser' }, body: 'My PR description', created_at: '2026-02-07T10:00:00Z' },
         { user: { login: 'maintainer' }, body: 'Thanks, will review soon', created_at: '2026-02-07T12:00:00Z' },
@@ -1494,7 +1541,8 @@ describe('isAcknowledgmentComment (Issue #69)', () => {
   });
 
   it('should still trigger hasUnrespondedComment for actionable comment after acknowledgment', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [
         { user: { login: 'testuser' }, body: 'My PR', created_at: '2026-02-07T10:00:00Z' },
         { user: { login: 'maintainer' }, body: 'Thanks, will look at this', created_at: '2026-02-07T11:00:00Z' },
@@ -1513,7 +1561,8 @@ describe('isAcknowledgmentComment (Issue #69)', () => {
   });
 
   it('should detect inline-only COMMENTED reviews as unresponded (#151)', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [{ user: { login: 'testuser' }, body: 'My PR', created_at: '2026-02-07T10:00:00Z' }],
       [
         // Review with inline comments only (no top-level body)
@@ -1527,7 +1576,8 @@ describe('isAcknowledgmentComment (Issue #69)', () => {
   });
 
   it('should skip reviews with null/undefined state and no body (#151)', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [{ user: { login: 'testuser' }, body: 'My PR', created_at: '2026-02-07T10:00:00Z' }],
       [{ user: { login: 'reviewer' }, body: '', submitted_at: '2026-02-07T12:00:00Z' }],
       [],
@@ -1537,7 +1587,8 @@ describe('isAcknowledgmentComment (Issue #69)', () => {
   });
 
   it('should skip CHANGES_REQUESTED reviews with empty body (#151)', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [{ user: { login: 'testuser' }, body: 'My PR', created_at: '2026-02-07T10:00:00Z' }],
       [{ user: { login: 'reviewer' }, body: '', submitted_at: '2026-02-07T12:00:00Z', state: 'CHANGES_REQUESTED' }],
       [],
@@ -1547,7 +1598,8 @@ describe('isAcknowledgmentComment (Issue #69)', () => {
   });
 
   it('should ignore user own COMMENTED reviews (#151)', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [],
       [
         // User's own inline review — should not trigger needs_response
@@ -1560,7 +1612,8 @@ describe('isAcknowledgmentComment (Issue #69)', () => {
   });
 
   it('should use synthetic body for inline-only COMMENTED reviews (#151)', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [],
       [{ user: { login: 'reviewer' }, body: '', submitted_at: '2026-02-07T12:00:00Z', state: 'COMMENTED' }],
       [],
@@ -1570,7 +1623,8 @@ describe('isAcknowledgmentComment (Issue #69)', () => {
   });
 
   it('should still skip APPROVED reviews with no body (#151)', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [{ user: { login: 'testuser' }, body: 'My PR', created_at: '2026-02-07T10:00:00Z' }],
       [{ user: { login: 'reviewer' }, body: '', submitted_at: '2026-02-07T12:00:00Z', state: 'APPROVED' }],
       [],
@@ -1580,8 +1634,9 @@ describe('isAcknowledgmentComment (Issue #69)', () => {
   });
 
   it('should detect new inline review after commit as needs_response (#151)', () => {
+    const monitor = new PRMonitor('fake-token');
     // Scenario: maintainer comments → user pushes commit → different reviewer posts inline review
-    const result = checkUnrespondedComments(
+    const result = (monitor as any).checkUnrespondedComments(
       [
         { user: { login: 'testuser' }, body: 'My PR', created_at: '2026-02-13T00:00:00Z' },
         { user: { login: 'fregante' }, body: 'Please fix X', created_at: '2026-02-13T00:37:25Z' },
@@ -2435,9 +2490,10 @@ describe('Maintainer self-reply detection (#199)', () => {
   });
 
   it('should skip COMMENTED review when all inline comments are self-replies', () => {
+    const monitor = new PRMonitor('fake-token');
     // Scenario from issue #199: MikeMcQuaid submits review, contributor pushes commit,
     // MikeMcQuaid self-replies to own inline comment
-    const result = checkUnrespondedComments(
+    const result = (monitor as any).checkUnrespondedComments(
       [{ user: { login: 'testuser' }, body: 'My PR', created_at: '2026-02-22T10:00:00Z' }],
       [
         // Original review with inline comments
@@ -2476,7 +2532,8 @@ describe('Maintainer self-reply detection (#199)', () => {
   });
 
   it('should not flag needs_response when only self-reply exists after user comment', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [
         { user: { login: 'testuser' }, body: 'My PR', created_at: '2026-02-22T10:00:00Z' },
         // User comments again after original review
@@ -2515,7 +2572,8 @@ describe('Maintainer self-reply detection (#199)', () => {
   });
 
   it('should still flag when maintainer replies to DIFFERENT author comment', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [{ user: { login: 'testuser' }, body: 'My PR', created_at: '2026-02-22T10:00:00Z' }],
       [
         // Review from reviewer-A
@@ -2550,7 +2608,8 @@ describe('Maintainer self-reply detection (#199)', () => {
   });
 
   it('should still flag when review has mix of self-reply and new thread', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [{ user: { login: 'testuser' }, body: 'My PR', created_at: '2026-02-22T10:00:00Z' }],
       [
         // Original review
@@ -2594,7 +2653,8 @@ describe('Maintainer self-reply detection (#199)', () => {
   });
 
   it('should use actual inline text instead of synthetic placeholder', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [],
       [
         // COMMENTED review with no body and no review.id → falls back to synthetic
@@ -2609,7 +2669,8 @@ describe('Maintainer self-reply detection (#199)', () => {
   });
 
   it('should use actual inline comment text when review has id and reviewComments exist', () => {
-    const result = checkUnrespondedComments(
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).checkUnrespondedComments(
       [],
       [{ user: { login: 'reviewer' }, body: '', submitted_at: '2026-02-07T12:00:00Z', state: 'COMMENTED', id: 100 }],
       [
@@ -2629,61 +2690,39 @@ describe('Maintainer self-reply detection (#199)', () => {
   });
 
   it('should return false from isAllSelfReplies when review has no inline comments', () => {
-    const result = isAllSelfReplies(999, []);
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).isAllSelfReplies(999, []);
     expect(result).toBe(false);
   });
 
   it('should return false from isAllSelfReplies when comment starts a new thread', () => {
-    const result = isAllSelfReplies(100, [
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).isAllSelfReplies(100, [
       // New thread (no in_reply_to_id)
-      {
-        id: 1000,
-        user: { login: 'maintainer' },
-        body: 'New issue found',
-        created_at: '2026-02-07T10:00:00Z',
-        pull_request_review_id: 100,
-      },
+      { id: 1000, user: { login: 'maintainer' }, body: 'New issue found', pull_request_review_id: 100 },
     ]);
     expect(result).toBe(false);
   });
 
   it('should return true from isAllSelfReplies when all comments reply to same author', () => {
-    const result = isAllSelfReplies(200, [
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).isAllSelfReplies(200, [
       // Parent comment from review 100
-      {
-        id: 1000,
-        user: { login: 'maintainer' },
-        body: 'Original comment',
-        created_at: '2026-02-07T10:00:00Z',
-        pull_request_review_id: 100,
-      },
+      { id: 1000, user: { login: 'maintainer' }, body: 'Original comment', pull_request_review_id: 100 },
       // Self-reply in review 200
-      {
-        id: 1001,
-        user: { login: 'maintainer' },
-        body: 'Follow up',
-        created_at: '2026-02-07T11:00:00Z',
-        in_reply_to_id: 1000,
-        pull_request_review_id: 200,
-      },
+      { id: 1001, user: { login: 'maintainer' }, body: 'Follow up', in_reply_to_id: 1000, pull_request_review_id: 200 },
     ]);
     expect(result).toBe(true);
   });
 
   it('should be case-insensitive when checking self-reply authors', () => {
-    const result = isAllSelfReplies(200, [
-      {
-        id: 1000,
-        user: { login: 'MikeMcQuaid' },
-        body: 'Original',
-        created_at: '2026-02-07T10:00:00Z',
-        pull_request_review_id: 100,
-      },
+    const monitor = new PRMonitor('fake-token');
+    const result = (monitor as any).isAllSelfReplies(200, [
+      { id: 1000, user: { login: 'MikeMcQuaid' }, body: 'Original', pull_request_review_id: 100 },
       {
         id: 1001,
         user: { login: 'mikemcquaid' },
         body: 'Follow up',
-        created_at: '2026-02-07T11:00:00Z',
         in_reply_to_id: 1000,
         pull_request_review_id: 200,
       },
